detect fs; fixed divide32 and remainder32; 32-bit itoa

+++ kernel/itoa.h
@@ -4,4 +4,4 @@
 
 extern char itoa_buf[];
 
-char* itoa(size_t num, char* buffer, uint8_t base);
+char* itoa(uint32_t num, char* buffer, uint8_t base);
